Crate libsavestate

A tiny crate to help emulators implement save states.

Note that this is wildly unsafe, but that's okay because it's open source.

Macros

impl_save_state!
Generates an impl of SaveState for a given type, saving/restoring a list of fields, and ignoring a second list of fields.

impl_save_state_fns!
Generates the functions of the SaveState trait.

impl_save_state_for_newtype!
Generates a SaveState impl for a newtype wrapper. The type must be declared as SomeType(InnerType) and InnerType must implement SaveState.

Traits

SaveState
Trait for types which can be serialized in a save state.

TransmuteByteSafe
Declares that a type can be safely transmuted into a byte slice of same length as the type's size in Bytes and vice-versa.

Functions

read_exact
Copied from std.