Trait libsavestate::SaveState

pub trait SaveState {
    fn save_state<W: Write + ?Sized>(&self, w: &mut W) -> Result<()>;
    fn restore_state<R: Read + ?Sized>(&mut self, r: &mut R) -> Result<()>;
}

Trait for types which can be serialized in a save state.

Note that this approach is a bit different from usual serialization: restore_state takes a &mut self and overwrites parts of it with the saved state. This way, we neither save the renderer, nor the ROM, nor the input sources, nor temporary data that is recomputed when we render a new frame after loading a save state.

When implementing this trait, use an exhaustive let-binding for all fields on the Self type. This way, any change that could potentially require a change in the SaveState implementation to the type will cause a compilation error. By using the impl_save_state! macro defined below, this will be done automatically and produce an error when the struct is changed.

Required Methods

fn save_state<W: Write + ?Sized>(&self, w: &mut W) -> Result<()>

fn restore_state<R: Read + ?Sized>(&mut self, r: &mut R) -> Result<()>