Trait libsavestate::TransmuteByteSafe

pub unsafe trait TransmuteByteSafe { }

Declares that a type can be safely transmuted into a byte slice of same length as the type's size in Bytes and vice-versa.

This is true for all types which don't impose any restrictions on their raw content, so this would not be safe to implement for types like bool or char. This is also obviously unsafe for non-#[repr(C)]-structs.